Move the policies to `h.security`
