--- h/auth/policy/__init__.py
@@ -1,2 +1,0 @@
-from h.auth.policy.bearer_token import TokenPolicy
-from h.auth.policy.combined import SecurityPolicy
