+++ h/security/policy/__init__.py
@@ -1,0 +1,2 @@
+from h.security.policy.bearer_token import TokenPolicy
+from h.security.policy.combined import SecurityPolicy
