+++ h/security/policy/_cookie.py
@@ -1,7 +1,7 @@
 from functools import lru_cache
 
-from h.auth.policy._identity_base import IdentityBasedPolicy
 from h.security import Identity
+from h.security.policy._identity_base import IdentityBasedPolicy
 from h.services import AuthCookieService
