+++ h/security/policy/combined.py
@@ -1,11 +1,11 @@
 from pyramid.interfaces import ISecurityPolicy
 from zope.interface import implementer
 
-from h.auth.policy._basic_http_auth import AuthClientPolicy
-from h.auth.policy._cookie import CookiePolicy
-from h.auth.policy._identity_base import IdentityBasedPolicy
-from h.auth.policy._remote_user import RemoteUserPolicy
-from h.auth.policy.bearer_token import TokenPolicy
+from h.security.policy._basic_http_auth import AuthClientPolicy
+from h.security.policy._cookie import CookiePolicy
+from h.security.policy._identity_base import IdentityBasedPolicy
+from h.security.policy._remote_user import RemoteUserPolicy
+from h.security.policy.bearer_token import TokenPolicy
 
 
 @implementer(ISecurityPolicy)
