+++ h/streamer/app.py
@@ -2,8 +2,8 @@ import os
 
 import pyramid
 
-from h.auth import TokenPolicy
 from h.config import configure
+from h.security import TokenPolicy
 from h.sentry_filters import SENTRY_FILTERS
 from h.streamer.worker import log
 
@@ -26,6 +26,7 @@ def create_app(_global_config, **settings):
     config.include("pyramid_services")
 
     config.include("h.auth")
+    config.include("h.security")
     # Override the default authentication policy.
     config.set_security_policy(TokenPolicy())
